AI_Programs/privacy_layer/requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24
pywaggle[vision]>=0.55